Ajustes na camada de captura e readme
